+++ doc/changelog.rst
@@ -90,7 +90,7 @@ PyMongo 4.2 brings a number of improvements including:
 - Support for the Queryable Encryption beta with MongoDB 6.0. Note that backwards-breaking
   changes may be made before the final release.  See :ref:`automatic-queryable-client-side-encryption` for example usage.
 - Provisional (beta) support for :func:`pymongo.timeout` to apply a single timeout
-  to an entire block of pymongo operations.
+  to an entire block of pymongo operations. See :ref:`timeout-example` for examples.
 - Added the ``timeoutMS`` URI and keyword argument to :class:`~pymongo.mongo_client.MongoClient`.
 - Added the :attr:`pymongo.errors.PyMongoError.timeout` property which is ``True`` when
   the error was caused by a timeout.